+++ core/src/main/java/net/automatalib/automata/vpda/AbstractOneSEVPA.java
@@ -125,7 +125,7 @@ public abstract class AbstractOneSEVPA<L, I> implements OneSEVPA<L, I>, Graph<L,
             for (final L loc : getLocations()) {
                 for (final I stackSymbol : alphabet.getCallAlphabet()) {
                     final int sym = encodeStackSym(loc, stackSymbol);
-                    final L succ = getReturnSuccessor(loc, i, sym);
+                    final L succ = getReturnSuccessor(location, i, sym);
 
                     if (succ != null) {
                         result.add(new SevpaViewEdge<>(i, sym, succ));
@@ -195,9 +195,9 @@ public abstract class AbstractOneSEVPA<L, I> implements OneSEVPA<L, I>, Graph<L,
 
     static class SevpaViewEdge<S, I> {
 
-        private final I input;
-        private final int stack;
-        private final S target;
+        final I input;
+        final int stack;
+        final S target;
 
         SevpaViewEdge(I input, int stack, S target) {
             this.target = target;